Sort category beers
